Move EnterpriseSnackbar
