+++ site/src/components/GlobalSnackbar/GlobalSnackbar.tsx
@@ -2,7 +2,7 @@ import { makeStyles } from "@mui/styles"
 import { useCallback, useState, FC } from "react"
 import { useCustomEvent } from "../../hooks/events"
 import { CustomEventListener } from "../../utils/events"
-import { EnterpriseSnackbar } from "../EnterpriseSnackbar/EnterpriseSnackbar"
+import { EnterpriseSnackbar } from "./EnterpriseSnackbar/EnterpriseSnackbar"
 import { ErrorIcon } from "../Icons/ErrorIcon"
 import { Typography } from "../Typography/Typography"
 import {
